doc: rename path to db file
